backend/requirements.txt
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4